Moved ryml header into seperate folder so it can be excluded from clang-formatting checks
