+++ src/include/simeng/yaml/.clang-format
@@ -1,0 +1,2 @@
+DisableFormat: true
+SortIncludes: Never
